@@ -1,31 +1,48 @@
 {
   "name": "@egovernments/digit-ui-module-core",
-  "version": "1.9.10",
+  "version": "2.0.0-dev-06",
   "license": "MIT",
-  "main": "dist/index.js",
-  "module": "dist/index.modern.js",
+  "main": "dist/main.js",
+  "module": "dist/main.js",
   "source": "src/Module.js",
   "files": [
     "dist"
   ],
   "scripts": {
-    "start": "microbundle-crl watch --no-compress --format modern,cjs",
-    "build": "microbundle-crl --compress --no-sourcemap --format cjs",
-    "prepublish": "yarn build"
+    "build": "cross-env NODE_ENV=production webpack --config webpack.config.js",
+    "build:dev": "cross-env NODE_ENV=development webpack --config webpack.config.js",
+    "build:analyze": "NODE_ENV=production webpack --config webpack.config.js --analyze",
+    "publish:components": "npm publish --tag console-v0.5"
+  },
+  "peerDependencies": {
+    "@tanstack/react-query": "^5.62.16",
+    "react": "19.0.0",
+    "react-dom": "19.0.0",
+    "react-router-dom": "6.25.1"
   },
   "dependencies": {
-    "@egovernments/digit-ui-components": "0.2.3",
-    "@egovernments/digit-ui-react-components": "1.9.0",
-    "@egovernments/digit-ui-svg-components": "1.1.0",
-    "react": "17.0.2",
-    "react-dom": "17.0.2",
-    "react-i18next": "11.16.2",
-    "react-query": "3.6.1",
+    "@egovernments/digit-ui-components": "2.0.0-dev-24",
+    "@egovernments/digit-ui-react-components": "2.0.0-dev-04",
+    "@egovernments/digit-ui-libraries": "2.0.0-dev-08",
+    "react-i18next": "15.0.0",
     "react-redux": "7.2.8",
-    "react-router-dom": "5.3.0",
+    "react-responsive": "10.0.1",
     "react-tooltip": "4.1.2",
-    "redux": "4.1.2",
+    "redux": "5.0.0",
     "redux-thunk": "2.4.1"
+  },
+  "sideEffects": false,
+  "browserslist": {
+    "production": [
+      ">0.2%",
+      "not dead",
+      "not op_mini all"
+    ],
+    "development": [
+      "last 1 chrome version",
+      "last 1 firefox version",
+      "last 1 safari version"
+    ]
   },
   "author": "JaganKumar <jagan.kumar@egov.org.in>",
   "keywords": [
@@ -34,5 +51,26 @@
     "dpg",
     "digit-ui",
     "core"
-  ]
+  ],
+  "devDependencies": {
+    "@babel/core": "^7.23.3",
+    "@babel/preset-env": "^7.23.3",
+    "@babel/preset-react": "^7.23.3",
+    "@tanstack/react-query": "^5.62.16",
+    "babel-loader": "^9.1.3",
+    "babel-plugin-transform-remove-console": "^6.9.4",
+    "core-js": "^3.33.0",
+    "cross-env": "7.0.3",
+    "css-loader": "^6.8.1",
+    "lint-staged": "12.3.7",
+    "react": "19.0.0",
+    "react-dom": "19.0.0",
+    "react-router-dom": "6.25.1",
+    "sass": "^1.83.4",
+    "sass-loader": "^16.0.4",
+    "style-loader": "^3.3.3",
+    "webpack": "^5.97.1",
+    "webpack-cli": "^5.1.4",
+    "webpack-dev-server": "^4.15.1"
+  }
 }
