@@ -1,11 +1,11 @@
 import React, {useRef, useEffect, useState } from "react";
 import { CheckCircle } from "@egovernments/digit-ui-svg-components";
-import { useHistory, useLocation } from "react-router-dom";
+import { useNavigate, useLocation } from "react-router-dom";
 import { useTranslation } from "react-i18next";
 
 const DummyLoaderScreen = () => {
   const [currentStep, setCurrentStep] = useState(0);
-  const history = useHistory();
+  const navigate = useNavigate();
   const location = useLocation();
   const { t } = useTranslation();
   const { tenant } = location.state || {};
@@ -29,11 +29,20 @@ const DummyLoaderScreen = () => {
 
     if (currentStep === steps.length) {
       clearInterval(stepInterval); // Clear the interval to stop further updates
+       const globalPath = typeof window !== 'undefined' ? window?.globalPath : '';
+
       const navigateTimeout = setTimeout(() => {
         if (roleForLandingPage(getUserRoles, MdmsRes)) {
-          window.location.href = `/${window?.globalPath}/${tenant}${RoleLandingUrl}`;
+        navigate({
+          pathname: `/${window?.globalPath}/${tenant}${RoleLandingUrl}`,
+          state: { tenant: tenant },
+        });
+
         } else {
-          window.location.href = `/${window?.globalPath}/${tenant}/employee`;
+          navigate({
+          pathname: `/${window?.globalPath}/${tenant}/employee`,
+          state: { tenant: tenant },
+        });
         }
       }, 1000);
 
