@@ -18,7 +18,7 @@ import {
 } from "@egovernments/digit-ui-react-components";
 import ReactTooltip from "react-tooltip";
 import { set } from "lodash";
-import { useHistory, useLocation, Link } from "react-router-dom";
+import { useNavigate, useLocation, Link } from "react-router-dom";
 import { useTranslation } from "react-i18next";
 
 
@@ -58,19 +58,19 @@ const findKey = (key = "") => {
 /*
 Used to navigate to other mission's ui if user has access
 */
-const navigateToRespectiveURL = (history = {}, url = "") => {
+const navigateToRespectiveURL = (navigate = {}, url = "") => {
   if (url?.indexOf(`/${window?.contextPath}`) === -1) {
     const hostUrl = window.location.origin;
     const updatedURL = hostUrl + url;
     window.location.href = updatedURL;
   } else {
-    history.push(url);
+    navigate(url);
   }
 };
 
 const Sidebar = ({ data }) => {
   const { t } = useTranslation();
-  const history = useHistory();
+  const navigate = useNavigate();
   const [openItems, setOpenItems] = useState({});
   const [selectedParent, setSelectedParent] = useState(null);
   const [selectedChild, setSelectedChild] = useState(null);
@@ -320,7 +320,7 @@ const Sidebar = ({ data }) => {
                   // setOpenItems({});
                   // setSelectedChildLevelOne(null)
                   navigateToRespectiveURL(
-                    history,
+                    navigate,
                     `${subItems?.item?.navigationURL}`
                   );
                 }}
