@@ -1,5 +1,5 @@
 import React, { useState, useContext } from "react";
-import { useHistory, useLocation, Link } from "react-router-dom";
+import { useNavigate, useLocation, Link } from "react-router-dom";
 import {
   ArrowForward,
   ArrowVectorDown,
@@ -25,7 +25,7 @@ const SubMenu = ({ item }) => {
   const location = useLocation();
   const { pathname } = location;
   const { t } = useTranslation();
-  const history = useHistory();
+  const navigate = useNavigate();
 
   const showSubnav = () => setSubnav(!subnav);
   const IconsObject = {
@@ -60,7 +60,7 @@ const SubMenu = ({ item }) => {
       <div className="submenu-container">
         <div className={`sidebar-link  ${pathname === item?.navigationURL ? "active" : ""}`}>
           <div className="actions">
-            <span style={{ marginLeft: "0px" }} onClick={() => history.push(`${item.navigationURL}`)}>
+            <span style={{ marginLeft: "0px" }} onClick={() => navigate(`${item.navigationURL}`)}>
               {leftIcon}
             </span>
             {item.navigationURL?.indexOf(`/${window?.contextPath}`) === -1 ? (
