@@ -1,6 +1,6 @@
 import React, { useState } from "react";
 import TopBar from "./TopBar";
-import { useHistory } from "react-router-dom";
+import { useNavigate } from "react-router-dom";
 import SideBar from "./SideBar";
 import LogoutDialog from "../Dialog/LogoutDialog";
 const TopBarSideBar = ({
@@ -19,7 +19,7 @@ const TopBarSideBar = ({
   islinkDataLoading,
 }) => {
   const [isSidebarOpen, toggleSidebar] = useState(false);
-  const history = useHistory();
+  const navigate = useNavigate()
   const [showDialog, setShowDialog] = useState(false);
   const handleLogout = () => {
     toggleSidebar(false);
@@ -37,7 +37,7 @@ const TopBarSideBar = ({
     toggleSidebar(!isSidebarOpen);
   };
   const userProfile = () => {
-    CITIZEN ? history.push(`/${window?.contextPath}/citizen/user/profile`) : history.push(`/${window?.contextPath}/employee/user/profile`);
+    CITIZEN ? navigate(`/${window?.contextPath}/citizen/user/profile`) : navigate(`/${window?.contextPath}/employee/user/profile`);
   };
   const userOptions = [
     { name: t("EDIT_PROFILE"), icon: "Edit", func: userProfile },
