@@ -12,12 +12,12 @@ import {
 } from "@egovernments/digit-ui-react-components";
 import React from "react";
 import { useTranslation } from "react-i18next";
-import { useHistory } from "react-router-dom";
+import { useNavigate } from "react-router-dom";
 import ImageComponent from "../../../components/ImageComponent";
 
 const Home = () => {
   const { t } = useTranslation();
-  const history = useHistory();
+  const navigate = useNavigate();
   const tenantId = Digit.Utils.getMultiRootTenant() ? Digit.ULBService.getStateId() : Digit.ULBService.getCitizenCurrentTenant(true);
   const { data: { stateInfo, uiHomePage } = {}, isLoading } = Digit.Hooks.useStore.getInitData();
   let isMobile = window.Digit.Utils.browser.isMobile();
@@ -37,7 +37,7 @@ const Home = () => {
   });
 
   if (!tenantId) {
-    history.push(`/${window?.contextPath}/citizen/select-language`);
+    navigate(`/${window?.contextPath}/citizen/select-language`);
   }
 
   const appBannerWebObj = uiHomePage?.appBannerDesktop;
@@ -50,11 +50,11 @@ const Home = () => {
   const redirectURL = uiHomePage?.redirectURL;
   /* configure redirect URL only if it is required to overide the default citizen home screen */
   if (redirectURL) {
-    history.push(`/${window?.contextPath}/citizen/${redirectURL}`);
+    navigate(`/${window?.contextPath}/citizen/${redirectURL}`);
   }
   /* fix for sanitation ui & sandbox*/
   if (window?.location?.href?.includes?.("sanitation-ui") || window?.location?.href?.includes?.("sandbox-ui")) {
-    history.push(`/${window?.contextPath}/citizen/all-services`);
+    navigate(`/${window?.contextPath}/citizen/all-services`);
   }
 
   const handleClickOnWhatsAppBanner = (obj) => {
@@ -65,23 +65,27 @@ const Home = () => {
     header: t(citizenServicesObj?.headerLabel),
     sideOption: {
       name: t(citizenServicesObj?.sideOption?.name),
-      onClick: () => history.push(citizenServicesObj?.sideOption?.navigationUrl),
+      onClick: () =>
+        navigate(citizenServicesObj?.sideOption?.navigationUrl),
     },
     options: [
       {
         name: t(citizenServicesObj?.props?.[0]?.label),
         Icon: <ComplaintIcon />,
-        onClick: () => history.push(citizenServicesObj?.props?.[0]?.navigationUrl),
+        onClick: () =>
+          navigate(citizenServicesObj?.props?.[0]?.navigationUrl),
       },
       {
         name: t(citizenServicesObj?.props?.[1]?.label),
         Icon: <PTIcon className="fill-path-primary-main" />,
-        onClick: () => history.push(citizenServicesObj?.props?.[1]?.navigationUrl),
+        onClick: () =>
+          navigate(citizenServicesObj?.props?.[1]?.navigationUrl),
       },
       {
         name: t(citizenServicesObj?.props?.[2]?.label),
         Icon: <CaseIcon className="fill-path-primary-main" />,
-        onClick: () => history.push(citizenServicesObj?.props?.[2]?.navigationUrl),
+        onClick: () =>
+          navigate(citizenServicesObj?.props?.[2]?.navigationUrl),
       },
       // {
       //     name: t("ACTION_TEST_WATER_AND_SEWERAGE"),
@@ -91,7 +95,8 @@ const Home = () => {
       {
         name: t(citizenServicesObj?.props?.[3]?.label),
         Icon: <OBPSIcon />,
-        onClick: () => history.push(citizenServicesObj?.props?.[3]?.navigationUrl),
+        onClick: () =>
+          navigate(citizenServicesObj?.props?.[3]?.navigationUrl),
       },
     ],
     styles: {
@@ -105,28 +110,32 @@ const Home = () => {
     header: t(infoAndUpdatesObj?.headerLabel),
     sideOption: {
       name: t(infoAndUpdatesObj?.sideOption?.name),
-      onClick: () => history.push(infoAndUpdatesObj?.sideOption?.navigationUrl),
+      onClick: () => navigate(infoAndUpdatesObj?.sideOption?.navigationUrl),
     },
     options: [
       {
         name: t(infoAndUpdatesObj?.props?.[0]?.label),
         Icon: <HomeIcon />,
-        onClick: () => history.push(infoAndUpdatesObj?.props?.[0]?.navigationUrl),
+        onClick: () =>
+          navigate(infoAndUpdatesObj?.props?.[0]?.navigationUrl),
       },
       {
         name: t(infoAndUpdatesObj?.props?.[1]?.label),
         Icon: <Calender />,
-        onClick: () => history.push(infoAndUpdatesObj?.props?.[1]?.navigationUrl),
+        onClick: () =>
+          navigate(infoAndUpdatesObj?.props?.[1]?.navigationUrl),
       },
       {
         name: t(infoAndUpdatesObj?.props?.[2]?.label),
         Icon: <DocumentIcon />,
-        onClick: () => history.push(infoAndUpdatesObj?.props?.[2]?.navigationUrl),
+        onClick: () =>
+          navigate(infoAndUpdatesObj?.props?.[2]?.navigationUrl),
       },
       {
         name: t(infoAndUpdatesObj?.props?.[3]?.label),
         Icon: <DocumentIcon />,
-        onClick: () => history.push(infoAndUpdatesObj?.props?.[3]?.navigationUrl),
+        onClick: () =>
+          navigate(infoAndUpdatesObj?.props?.[3]?.navigationUrl),
       },
       // {
       //     name: t("CS_COMMON_HELP"),
@@ -191,7 +200,7 @@ const Home = () => {
             <div className="WhatsNewSection">
               <div className="headSection">
                 <h2>{t(whatsNewSectionObj?.headerLabel)}</h2>
-                <p onClick={() => history.push(whatsNewSectionObj?.sideOption?.navigationUrl)}>{t(whatsNewSectionObj?.sideOption?.name)}</p>
+                <p onClick={() => navigate(whatsNewSectionObj?.sideOption?.navigationUrl)}>{t(whatsNewSectionObj?.sideOption?.name)}</p>
               </div>
               <WhatsNewCard {...EventsData?.[0]} />
             </div>
