@@ -2,7 +2,7 @@ import { CardSubHeader, FormComposer,CardText} from "@egovernments/digit-ui-reac
 import { BackLink,Toast} from "@egovernments/digit-ui-components";
 import PropTypes from "prop-types";
 import React, { useEffect, useState } from "react";
-import { useHistory } from "react-router-dom";
+import { useNavigate } from "react-router-dom";
 import Background from "../../../components/Background";
 import Header from "../../../components/Header";
 import SelectOtp from "../../citizen/Login/SelectOtp";
@@ -11,7 +11,7 @@ import ImageComponent from "../../../components/ImageComponent";
 const ChangePasswordComponent = ({ config: propsConfig, t }) => {
   const [user, setUser] = useState(null);
   const { mobile_number: mobileNumber, tenantId } = Digit.Hooks.useQueryParams();
-  const history = useHistory();
+  const navigate = useNavigate();
   const [otp, setOtp] = useState("");
   const [isOtpValid, setIsOtpValid] = useState(true);
   const [showToast, setShowToast] = useState(null);
@@ -23,7 +23,7 @@ const ChangePasswordComponent = ({ config: propsConfig, t }) => {
     }
     Digit.UserService.setUser(user);
     const redirectPath = location.state?.from || `/${window?.contextPath}/employee`;
-    history.replace(redirectPath);
+    navigate(redirectPath, {replace:true});
   }, [user]);
 
   const closeToast = () => {
@@ -70,7 +70,7 @@ const ChangePasswordComponent = ({ config: propsConfig, t }) => {
   };
 
   const navigateToLogin = () => {
-    history.replace(`/${window?.contextPath}/employee/user/login`);
+    navigate(`/${window?.contextPath}/employee/user/login`, {replace:true});
   };
 
   const [username, password, confirmPassword] = propsConfig.inputs;
