@@ -2,7 +2,7 @@
 import { BackLink, Dropdown, Loader, Toast, FormComposerV2 } from "@egovernments/digit-ui-components";
 import PropTypes from "prop-types";
 import React, { useEffect, useState } from "react";
-import { useHistory } from "react-router-dom";
+import { useNavigate } from "react-router-dom";
 import Background from "../../../components/Background";
 import Header from "../../../components/Header";
 import ImageComponent from "../../../components/ImageComponent";
@@ -12,7 +12,7 @@ import { useLoginConfig } from "../../../hooks/useLoginConfig";
 const ForgotPassword = ({ config: propsConfig, t, stateCode }) => {
   const { data: cities, isLoading } = Digit.Hooks.useTenants();
   const [user, setUser] = useState(null);
-  const history = useHistory();
+  const navigate = useNavigate();
   const [showToast, setShowToast] = useState(null);
   const getUserType = () => Digit.UserService.getType();
   
@@ -30,7 +30,7 @@ const ForgotPassword = ({ config: propsConfig, t, stateCode }) => {
     }
     Digit.UserService.setUser(user);
     const redirectPath = location.state?.from || `/${window?.contextPath}/employee`;
-    history.replace(redirectPath);
+    navigate(redirectPath, {replace:true});
   }, [user]);
 
   const closeToast = () => {
@@ -52,7 +52,7 @@ const ForgotPassword = ({ config: propsConfig, t, stateCode }) => {
     };
     try {
       await Digit.UserService.sendOtp(requestData, data.city.code);
-      history.push(`/${window?.contextPath}/employee/user/change-password?USERNAME=${data.username}&tenantId=${data.city.code}`);
+      navigate(`/${window?.contextPath}/employee/user/change-password?USERNAME=${data.username}&tenantId=${data.city.code}`);
     } catch (err) {
       setShowToast(err?.response?.data?.error?.fields?.[0]?.message || "Invalid login credentials!");
       setTimeout(closeToast, 5000);
@@ -60,7 +60,7 @@ const ForgotPassword = ({ config: propsConfig, t, stateCode }) => {
   };
 
   const navigateToLogin = () => {
-    history.replace(`/${window?.contextPath}/employee/login`);
+    navigate(`/${window?.contextPath}/employee/login`);
   };
 
   const [userId, city] = propsConfig.inputs;
