@@ -1,33 +1,31 @@
 import React, { useMemo } from "react";
 import { useTranslation } from "react-i18next";
-import { Route, Switch, useRouteMatch } from "react-router-dom";
+import { Routes, Route } from "react-router-dom";
 import { loginConfig } from "./config";
 import ForgotPasswordComponent from "./forgotPassword";
 
-const EmployeeForgotPassword = ({stateCode}) => {
+const EmployeeForgotPassword = ({ stateCode }) => {
   const { t } = useTranslation();
-  const { path } = useRouteMatch();
 
-
-  const params = useMemo(() =>
-    loginConfig.map(
-      (step) => {
+  const params = useMemo(
+    () =>
+      loginConfig.map((step) => {
         const texts = {};
         for (const key in step.texts) {
           texts[key] = t(step.texts[key]);
         }
         return { ...step, texts };
-      },
-      [loginConfig]
-    )
+      }),
+    [loginConfig]
   );
 
   return (
-    <Switch>
-      <Route path={`${path}`} exact>
-        <ForgotPasswordComponent config={params[0]}  t={t} stateCode={stateCode}/>
-      </Route>
-    </Switch>
+    <Routes>
+      <Route
+        path=""
+        element={<ForgotPasswordComponent config={params[0]} t={t} stateCode={stateCode} />}
+      />
+    </Routes>
   );
 };
 
