@@ -1,18 +1,17 @@
 import React, { useEffect, useMemo, useState } from "react";
 import { useTranslation } from "react-i18next";
-import { Route, Switch, useRouteMatch } from "react-router-dom";
-import { SignUpConfig as defaultSignUpConfig  } from "./config";
+import { Routes, Route, useLocation } from "react-router-dom";
+import { SignUpConfig as defaultSignUpConfig } from "./config";
 import Login from "./signUp";
-import { useHistory, useLocation } from "react-router-dom";
 
-
-const SignUp = ({stateCode}) => {
+const SignUp = ({ stateCode }) => {
   const { t } = useTranslation();
-  const { path } = useRouteMatch();
   const [SignUpConfig, setSignUpConfig] = useState(defaultSignUpConfig);
+
   const moduleCode = ["privacy-policy"];
   const language = Digit.StoreData.getCurrentLanguage();
   const modulePrefix = "digit";
+
   const { data: store } = Digit.Services.useStore({
     stateCode,
     moduleCode,
@@ -20,61 +19,45 @@ const SignUp = ({stateCode}) => {
     modulePrefix
   });
 
-  const history = useHistory();
-  const location = useLocation();
-
-
-    // Timestamp handling
-    useEffect(() => {
-      const query = new URLSearchParams(location.search);
-      if (!query.get("ts")) {
-        const ts = Date.now();
-        history.replace({
-          pathname: location.pathname,
-          search: `?ts=${ts}`
-        });
-      }
-    }, [location, history]);
-
-  const { data: mdmsData, isLoading } = Digit.Hooks.useCommonMDMS(stateCode, "commonUiConfig", ["SignUpConfig"], {
-    select: (data) => {
-      return {
+  const { data: mdmsData, isLoading } = Digit.Hooks.useCommonMDMS(
+    stateCode,
+    "commonUiConfig",
+    ["SignUpConfig"],
+    {
+      select: (data) => ({
         config: data?.commonUiConfig?.SignUpConfig
-      };
-    },
-    retry: false,
-  });
-
-  // let SignUpConfig = mdmsData?.config ? mdmsData?.config : defaultSignUpConfig;
-  useEffect(() => {
-    if(isLoading == false && mdmsData?.config)
-    {  
-      setSignUpConfig(mdmsData?.config)
-    }else{
-      setSignUpConfig(defaultSignUpConfig)
+      }),
+      retry: false
     }
-  },[mdmsData, isLoading])
+  );
 
+  useEffect(() => {
+    if (!isLoading && mdmsData?.config) {
+      setSignUpConfig(mdmsData.config);
+    } else {
+      setSignUpConfig(defaultSignUpConfig);
+    }
+  }, [mdmsData, isLoading]);
 
-  const SignUpParams = useMemo(() =>
-    SignUpConfig.map(
-      (step) => {
+  const SignUpParams = useMemo(
+    () =>
+      SignUpConfig.map((step) => {
         const texts = {};
         for (const key in step.texts) {
           texts[key] = t(step.texts[key]);
         }
         return { ...step, texts };
-      },
-      [SignUpConfig]
-    )
+      }),
+    [SignUpConfig]
   );
 
   return (
-    <Switch>
-      <Route path={`${path}`} exact>
-        <Login config={SignUpParams[0]} t={t} />
-      </Route>
-    </Switch>
+    <Routes>
+      <Route
+        path=""
+        element={<Login config={SignUpParams[0]} t={t} />}
+      />
+    </Routes>
   );
 };
 
