@@ -2,7 +2,6 @@ import { Loader } from "@egovernments/digit-ui-components";
 import React from "react";
 import { useTranslation } from "react-i18next";
 // import ImageComponent from "./ImageComponent";
-import { ReactComponent as SandboxHeaderImg } from '../components/Images/SandboxHeader.svg';
 
 const SandBoxHeader = ({ showTenant = true }) => {
   const { data: storeData, isLoading } = Digit.Hooks.useStore.getInitData();
@@ -14,7 +13,7 @@ const SandBoxHeader = ({ showTenant = true }) => {
   return (
     <div className="bannerHeader">
         <div className="SandboxLogoAlignCenter">
-        <SandboxHeaderImg/>
+        {/* <SandboxHeaderImg/> */}
         </div>
       {/* <ImageComponent className="bannerLogo" src={stateInfo?.logoUrl} style={!showTenant ? { borderRight: "unset" } : {}} alt="Digit Banner" /> */}
       {showTenant && stateInfo?.code && <p>{t(`TENANT_TENANTS_${stateInfo?.code?.toUpperCase()}`)}</p>}
