@@ -2,7 +2,7 @@ import React, { Fragment, useState } from "react";
 import { CardText, CardLabelError, Toast, CardLabel } from "@egovernments/digit-ui-components";
 import { OTPInput } from "@egovernments/digit-ui-react-components";
 import { useTranslation } from "react-i18next";
-import { useRouteMatch, useHistory, useLocation } from "react-router-dom";
+import { useLocation } from "react-router-dom";
 import useInterval from "../../../hooks/useInterval";
 import Background from "../../../components/Background";
 import { useEffect } from "react";
