@@ -11,7 +11,9 @@ const getRootReducer = (defaultStore, moduleReducers) =>
 const middleware = [thunk];
 
 const composeEnhancers =
-  typeof window === "object" && window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__ ? window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__({}) : compose;
+  process.env.NODE_ENV === 'development' && typeof window === "object" && window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__ 
+    ? window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__({}) 
+    : compose;
 
 const enhancer = composeEnhancers(
   applyMiddleware(...middleware)
